+++ src/test/components/MenuPanel.test.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { fireEvent } from '@testing-library/react';
-import renderConnected from '../utils/renderConnected';
+import renderConnected from '../test_utils/renderConnected';
 import MenuPanel from '../../components/MenuPanel';
 
 describe('<MenuPanel/>', () => {
